Edit an Item!
Comment out more phoenix.css as we don't need more of this.
Change form template to have a unique id as well as an .item-form class we can attach css to more easily.
Add @editing assign that captures the item we're editing, empty for index.
Add editing/2 function to enable the "editing" class for the li when editing.
Add toggle_edit event that when clicked will assign the item as @editing. This enables our editing form.
Change `.todo-list li label` css to bump up the padding so it's closer to the form size
Add `.view a, .view a:visited` to change our editing link so it doesn't look like a link.