Extra credit: Blur text
Use LiveView communication between component and parent on blur so that when focus is out of an edit box, editing is turned off.
This surprisingly took very little to find a solution through Google. I saw PubSub mentioned but I think component communication is more ideal. We may wish to add PubSub later so multiple tabs are kept in sync.