Fix tests
Strip out unused portions, though this could likely go deeper into the router. It would be useful to try to have better test coverage but I'm a little lost on what would make for good tests from here. The scaffolding is mostly gone and only retains a husk of what it once was.

I may or may not be done done now. We'll see if I want to futz with this again in the future or maybe start from scratch. It's not a bad project to work through rather quickly.