Screenshots cropped
Use a combination of https://www.batchphoto.com/blog/how-to-crop-multiple-images-at-once-in-photoshop/ and changing the Canvas Size for `todo-completed-list.png` as the url in the bottom left somehow adjusted this one screen outside the norm. Used the anchor point in the top middle so it looked something like `<- / * \ ->` where the dot was in the top middle section. We wanted to keep our top orientation and pull the canvas size in to match because I had cropped too much.

I may never remember this technique was described in this git commit message but there will likely be a time where I do this again.